+++ RocksmithToolkitLib/DLCPackage/Manifest2014/Tone/Pedal2014.cs
@@ -9,14 +9,14 @@ namespace RocksmithToolkitLib.DLCPackage.Manifest.Tone
     public class Pedal2014
     {
         public string Type { get; set; }
-        public Dictionary<string, decimal> KnobValues { get; set; }
+        public Dictionary<string, float> KnobValues { get; set; }
         [JsonProperty("Key")]
         public string PedalKey { get; set; }
         public string Category { get; set; }
 
         public Pedal2014()
         {
-            KnobValues = new Dictionary<string, decimal>();
+            KnobValues = new Dictionary<string, float>();
         }
 
         public bool Any() {
